+++ R7.University.EmployeeDirectory/ViewEmployeeDirectory.ascx.cs
@@ -4,7 +4,7 @@
 // Author:
 //       Roman M. Yagodin <roman.yagodin@gmail.com>
 //
-// Copyright (c) 2014 
+// Copyright (c) 2014-2015
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -33,6 +33,7 @@ using DotNetNuke.Common.Utilities;
 using DotNetNuke.Services.Exceptions;
 using DotNetNuke.Entities.Icons;
 using R7.University;
+using R7.University.Extensions;
 
 namespace R7.University.EmployeeDirectory
 {
