+++ R7.University/Extensions/GridViewExtensions.cs
@@ -29,7 +29,7 @@ using System.Web.UI.WebControls;
 using DotNetNuke.Entities.Modules;
 using DotNetNuke.Services.Localization;
 
-namespace R7.University
+namespace R7.University.Extensions
 {
     public static class GridViewExtensions
     {
